requirements.txt
Django>=4.0,<4.1
djangorestframework>=3.12,<3.13
psycopg2<=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16